+++ docs/CHANGES.md
@@ -1,1 +1,4 @@
-- Removed - Temporarily disabled Battle Pet logic - [BLU.toc]
+- Removed - Debug for missing localization - [localizaton.lua]
+- Removed - Excess debug output - [core.lua, initialization.lua, utils.lua]
+- Removed - (temp) Redacted - 'Battle-Pets Level-Up!' - [core.lua, initialization.lua, , options.lua, utils.lua]
+- Removed - Unused Dependancy - [BLU.toc, BLU_Vanilla.toc, BLU_Cata.toc]